+++ src/Telegram.Bot/ITelegramBotClient.cs
@@ -338,7 +338,7 @@ namespace Telegram.Bot
         /// <see href="https://core.telegram.org/bots/api#senddocument"/>
         Task<Message> SendDocumentAsync(
             ChatId chatId,
-            FileToSend document,
+            InputOnlineFile document,
             string caption = default,
             bool disableNotification = default,
             int replyToMessageId = default,